backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2
openai==1.12.0
requests==2.31.0
urllib3==2.2.1
certifi==2024.2.2
gunicorn==21.2.0
PyMuPDF==1.23.7
uuid==1.30
traceback2==1.4.0
